ci: run pybambu unit tests on every PR, plus the missing pytest dep fix
