+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-A1.json
