+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-A2L.json
