+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-H2C.json
