+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-P1PNOAMS.json
