+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-P2S.json
