+++ custom_components/bambu_lab/pybambu/mock_data/MOCK-X2D.json
