--- custom_components/bambu_lab/pybambu/tests/__init__.py
@@ -1,7 +1,0 @@
-"""
-Test package for pybambu library
-"""
-
-from .test_utils import MockMQTTClient
-
-__all__ = ['MockMQTTClient']
